Modified the gradle debian package
For simplifying the installation I added back the application.properties having some default configuration. In case no configuration file is passed, the default behaviour is to use the in-memory db (hsql).
in order to create a debian package it is necessary to execute:
$./gradlew openbatonDeb

you will find a debian package in the gradle/distributions/ folder. You can install this debian package using
$sudo dpkg -i openbaton.deb

and start openbaton
$openbaton

upstart-job is also available:
$service start openbaton

but I'm still having some issues to solve with it