Removed option of passing image inside VNFPackage
Since it is not possible anymore to pass the whole image file as an argument
when calling the add image method, it was decided to remove this option temporarly.
If a VNFPackage contains an image, the VNFPackage onboarding will be rejected
by throwing an Exception. In this way the user is notified and forced to use the image-link
defined in the Metadata.yaml.