Fix two bugs
First bug was that the VNFR was saved in the instantiate task before the parameters for its dependencies were filled. Resulting in missing parameters in scripts of other VNFR sometimes.
The second bug was that an exception thrown while saving a VNFRecordDependency to the database was not caught.
Furthermore if an exception occured while calling doWork() on a task like instantiate , modify etc. the VNFR's status was set to error but not saved.