Imp: Let the log file request use the same queue as all the other requests
Before it used nfvo.vnfm.logs. Furthermore VNFMs which do not implement the functionality to request logs will not throw an exception but return empty logs with an error log mentioning that this functionality is not supported. For this commit the LOG_REQUEST action was added and two new message types for sending the logs.