Fix: Deadlock when deploying many NSs in parallel
Reason for the deadlock was the threadpool. Now the default configuration does not limit the thread pool's size. If other values are set for corepoolsize and queuecapacity, a deadlock can occur again. Some changes were made to the VNFStateHandler and related components in order to reduce the number of started threads (without losing the concurrent execution).